advanced-api-project/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=22.1
